- ditch warnings - fix PHP4 build, yet again - http_message_parse() must check for Transfer-Encoding prior Content-Length

+++ php_http_cache_api.h
@@ -86,7 +86,7 @@ static inline void *_http_etag_init(TSRMLS_D)
 		
 #ifdef HAVE_LIBMHASH
 		default:
-			if ((mode < 0) || (mode > mhash_count()) || (!(ctx = mhash_init(mode)))) {
+			if ((mode < 0) || ((ulong)mode > mhash_count()) || (!(ctx = mhash_init(mode)))) {
 				http_error_ex(HE_ERROR, HE_RUNTIME, "Invalid ETag mode: %ld", mode);
 			}
 		break;
